Ask what divides, not only what the compiler will name
Two legs failed for two reasons, and the width ceiling had a second source.

clang-cl links neither compiler-rt nor its division builtins:

    lld-link : error : undefined symbol: __udivti3
    lld-link : error : undefined symbol: __umodti3

That is a link-time answer to a question __BITINT_MAXWIDTH__ cannot be asked:
the compiler will name a 128-bit type there, and then nothing divides it. So the
ceiling is now the lower of what the compiler names and what the target divides
by itself, and the lists are built in three tiers - through 64 everywhere, 128
where the ceiling allows, 256 above that. Verified in all four configurations:
upstream Clang gives 256, Apple's 128, and clang-cl 64 either way.

clang-tidy 22 found four things in the new code, all of them fair:

- readability-math-missing-parentheses on the quotient identity, three times
- misc-include-cleaner: bit_int, bit_uint and make_unsigned_like_t were used
  through headers the file did not include directly
- readability-use-concise-preprocessor-directives: #if defined(X) is #ifdef X
- readability-function-cognitive-complexity: sweep scored 31 against a 25
  threshold, and is now three named checks over one pair, which reads better
  than it did before the tool objected
- readability-implicit-bool-conversion where storage_width adds is_signed to
  digits, now a static_cast

And one comment stated a reason that is not one: a signed width of 1 does not
lack room for a magnitude, it holds -1 and 0, which is why WG14 N3644 proposes
allowing it. C23 setting the signed minimum at two bits is the whole reason the
signed list starts there, so that is what the comment now says.

+++ test/include/xstd/test/integer_class.hpp
@@ -25,7 +25,8 @@ struct storage_limits : std::numeric_limits<Storage>
 
 // Sign bit included, so that a counterpart is the same width rather than the same digits.
 template<class Storage>
-inline constexpr auto storage_width = storage_limits<Storage>::digits + storage_limits<Storage>::is_signed;
+inline constexpr auto storage_width =
+        storage_limits<Storage>::digits + static_cast<int>(storage_limits<Storage>::is_signed);
 
 // What the binary operators return in the proxy variant: convertible to the type, not it.
 template<class Storage, bool ProxyResults>
